Set up installed headers and pkg-config for libeuicc-drivers.so
Also separate lpac_driver struct from the installed version
