+++ driver/apdu/at.h
@@ -1,4 +1,4 @@
 #pragma once
-#include <driver.h>
+#include <driver.private.h>
 
 extern const struct lpac_driver driver_apdu_at;
