+++ driver/libeuicc-drivers.pc.in
@@ -1,0 +1,11 @@
+prefix="@CMAKE_INSTALL_PREFIX@"
+exec_prefix="${prefix}"
+libdir="${prefix}/lib"
+includedir="${prefix}/include"
+
+Name: libeuicc-drivers
+Description: An "official" collection of drivers (backends) and their loader for use with libeuicc
+Version: @PROJECT_VERSION@
+Requires: libeuicc = @PROJECT_VERSION@
+Cflags: -I${includedir}
+Libs: -L${libdir} -leuicc-drivers
